Split home refresh pipeline tests and fakes #293
