+++ Tests/UnitTests/HomeRefreshPipelineTests.swift
@@ -1829,29 +1829,6 @@ private actor SequencedHomeIngestionCoordinator: HomeIngestionCoordinating {
     }
 }
 
-@Suite("SkyAware App Activation Cleanup")
-struct SkyAwareAppActivationCleanupTests {
-    @Test("shouldRunActivationCleanup_enforcesHourlyThrottle")
-    func shouldRunActivationCleanupEnforcesHourlyThrottle() {
-        let now = Date(timeIntervalSinceReferenceDate: 2_000_000)
-        let minimumInterval = ActivationCleanupThrottle.minimumInterval
-
-        #expect(ActivationCleanupThrottle.shouldRun(lastRunAt: 0, now: now))
-        #expect(
-            ActivationCleanupThrottle.shouldRun(
-                lastRunAt: now.timeIntervalSinceReferenceDate - minimumInterval + 1,
-                now: now
-            ) == false
-        )
-        #expect(
-            ActivationCleanupThrottle.shouldRun(
-                lastRunAt: now.timeIntervalSinceReferenceDate - minimumInterval,
-                now: now
-            )
-        )
-    }
-}
-
 private actor AsyncGate {
     private var continuation: CheckedContinuation<Void, Never>?
     private var isOpen = false
